@@ -73,5 +14,103 @@
     "content",
     "management",
     "system"
-  ]
+  ],
+  "bugs": {
+    "url": "https://github.com/bolt/four/issues"
+  },
+  "main": "webpack.config.js",
+  "directories": {
+    "tests": "tests"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git://github.com/bolt/four.git"
+  },
+  "dependencies": {
+    "axios": "^0.18.0",
+    "baguettebox.js": "^1.11.0",
+    "bootstrap": "^4.2.1",
+    "codemirror": "^5.42.2",
+    "flagpack": "^1.0.3",
+    "jquery": "^3.3.1",
+    "moment": "^2.23.0",
+    "no-scroll": "^2.1.1",
+    "popper.js": "^1.14.6",
+    "simplemde": "^1.11.2",
+    "tinycolor2": "^1.4.1",
+    "vue": "^2.5.21",
+    "vue-flatpickr-component": "^8.1.1",
+    "vue-multiselect": "^2.1.3",
+    "vue-notification": "^1.3.14",
+    "vue-simplemde": "^0.5.1",
+    "vue-trumbowyg": "^3.4.1",
+    "vuedraggable": "^2.17.0",
+    "vuex": "^3.0.1"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.2.2",
+    "@babel/plugin-transform-regenerator": "^7.0.0",
+    "@babel/plugin-transform-runtime": "^7.2.0",
+    "@babel/polyfill": "^7.2.5",
+    "@babel/preset-env": "^7.2.3",
+    "@symfony/webpack-encore": "^0.22.4",
+    "@vue/test-utils": "^1.0.0-beta.28",
+    "ajv-keywords": "^3.2.0",
+    "autoprefixer": "^9.4.4",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-jest": "^23.6.0",
+    "eslint": "5.12.0",
+    "eslint-config-prettier": "^3.3.0",
+    "eslint-plugin-prettier": "^3.0.1",
+    "eslint-plugin-vue": "^5.1.0",
+    "jest": "^23.6.0",
+    "jest-serializer-vue": "^2.0.2",
+    "node-sass": "^4.11.0",
+    "postcss-loader": "^3.0.0",
+    "postcss-preset-env": "^6.5.0",
+    "prettier": "1.15.3",
+    "regenerator-runtime": "^0.13.1",
+    "register-service-worker": "^1.5.2",
+    "sass-loader": "^7.1.0",
+    "vue-jest": "^3.0.2",
+    "vue-loader": "^15.5.0",
+    "vue-router": "^3.0.2",
+    "vue-template-compiler": "^2.5.21",
+    "webpackbar": "^3.1.4",
+    "workbox-webpack-plugin": "^3.6.3"
+  },
+  "scripts": {
+    "lint": "eslint --ext .js,.vue, assets",
+    "lint-fix": "eslint --ext .js,.vue, assets --fix",
+    "serve": "encore dev-server --port=8234 --progress --disable-host-check",
+    "build": "encore production --progress",
+    "test": "jest"
+  },
+  "eslintConfig": {
+    "root": true,
+    "extends": [
+      "plugin:vue/essential",
+      "eslint:recommended"
+    ]
+  },
+  "browserslist": [
+    "> 1%"
+  ],
+  "jest": {
+    "moduleFileExtensions": [
+      "js",
+      "json",
+      "vue"
+    ],
+    "transform": {
+      "^.+\\.js$": "<rootDir>/node_modules/babel-jest",
+      ".*\\.(vue)$": "<rootDir>/node_modules/vue-jest"
+    },
+    "moduleNameMapper": {
+      "^@/(.*)$": "<rootDir>/assets/js/app/$1"
+    },
+    "snapshotSerializers": [
+      "<rootDir>/node_modules/jest-serializer-vue"
+    ]
+  }
 }
