@@ -14,7 +14,7 @@
                             <td><a :href="'edit/'+record.id">{{ record.fields[0].value.value }}</a></td>
                         </tr>
                         <!-- Maybe is better to have a component to print each row? -->
-                        <!-- <Context :id="item.id" :key="item.id" :contenttype="content.contenttype" :title="item.fields[0]"></Context> -->
+                        <!-- <Context :id="item.id" :key="item.id" :contentType="content.contentType" :title="item.fields[0]"></Context> -->
                     </template>
                 </tbody>
             </table>
