@@ -66,7 +66,7 @@ class AddUserCommand extends Command
     /**
      * {@inheritdoc}
      */
-    protected function configure()
+    protected function configure(): void
     {
         $this
             ->setDescription('Creates users and stores them in the database')
@@ -76,16 +76,15 @@ class AddUserCommand extends Command
             ->addArgument('username', InputArgument::OPTIONAL, 'The username of the new user')
             ->addArgument('password', InputArgument::OPTIONAL, 'The plain password of the new user')
             ->addArgument('email', InputArgument::OPTIONAL, 'The email of the new user')
-            ->addArgument('full-name', InputArgument::OPTIONAL, 'The full name of the new user')
-            ->addOption('admin', null, InputOption::VALUE_NONE, 'If set, the user is created as an administrator')
-        ;
+            ->addArgument('display-name', InputArgument::OPTIONAL, 'The display name of the new user')
+            ->addOption('admin', null, InputOption::VALUE_NONE, 'If set, the user is created as an administrator');
     }
 
     /**
      * This optional method is the first one executed for a command after configure()
      * and is useful to initialize properties based on the input arguments and options.
      */
-    protected function initialize(InputInterface $input, OutputInterface $output)
+    protected function initialize(InputInterface $input, OutputInterface $output): void
     {
         // SymfonyStyle is an optional feature that Symfony provides so you can
         // apply a consistent look to the commands of your application.
@@ -103,9 +102,9 @@ class AddUserCommand extends Command
      * quite a lot of work. However, if the command is meant to be used by external
      * users, this method is a nice way to fall back and prevent errors.
      */
-    protected function interact(InputInterface $input, OutputInterface $output)
+    protected function interact(InputInterface $input, OutputInterface $output): void
     {
-        if (null !== $input->getArgument('username') && null !== $input->getArgument('password') && null !== $input->getArgument('email') && null !== $input->getArgument('full-name')) {
+        if ($input->getArgument('username') !== null && $input->getArgument('password') !== null && $input->getArgument('email') !== null && $input->getArgument('display-name') !== null) {
             return;
         }
 
@@ -121,7 +120,7 @@ class AddUserCommand extends Command
 
         // Ask for the username if it's not defined
         $username = $input->getArgument('username');
-        if (null !== $username) {
+        if ($username !== null) {
             $this->io->text(' > <info>Username</info>: ' . $username);
         } else {
             $username = $this->io->ask('Username', null, [$this->validator, 'validateUsername']);
@@ -130,7 +129,7 @@ class AddUserCommand extends Command
 
         // Ask for the password if it's not defined
         $password = $input->getArgument('password');
-        if (null !== $password) {
+        if ($password !== null) {
             $this->io->text(' > <info>Password</info>: ' . str_repeat('*', mb_strlen($password)));
         } else {
             $password = $this->io->askHidden('Password (your type will be hidden)', [$this->validator, 'validatePassword']);
@@ -139,20 +138,20 @@ class AddUserCommand extends Command
 
         // Ask for the email if it's not defined
         $email = $input->getArgument('email');
-        if (null !== $email) {
+        if ($email !== null) {
             $this->io->text(' > <info>Email</info>: ' . $email);
         } else {
             $email = $this->io->ask('Email', null, [$this->validator, 'validateEmail']);
             $input->setArgument('email', $email);
         }
 
-        // Ask for the full name if it's not defined
-        $fullName = $input->getArgument('full-name');
-        if (null !== $fullName) {
-            $this->io->text(' > <info>Full Name</info>: ' . $fullName);
+        // Ask for the display name if it's not defined
+        $displayName = $input->getArgument('display-name');
+        if ($displayName !== null) {
+            $this->io->text(' > <info>Display Name</info>: ' . $displayName);
         } else {
-            $fullName = $this->io->ask('Full Name', null, [$this->validator, 'validateFullName']);
-            $input->setArgument('full-name', $fullName);
+            $displayName = $this->io->ask('Display Name', null, [$this->validator, 'validateDisplayName']);
+            $input->setArgument('display-name', $displayName);
         }
     }
 
@@ -168,15 +167,15 @@ class AddUserCommand extends Command
         $username = $input->getArgument('username');
         $plainPassword = $input->getArgument('password');
         $email = $input->getArgument('email');
-        $fullName = $input->getArgument('full-name');
+        $displayName = $input->getArgument('display-name');
         $isAdmin = $input->getOption('admin');
 
         // make sure to validate the user data is correct
-        $this->validateUserData($username, $plainPassword, $email, $fullName);
+        $this->validateUserData($username, $plainPassword, $email, $displayName);
 
         // create the user and encode its password
         $user = new User();
-        $user->setFullName($fullName);
+        $user->setDisplayName($displayName);
         $user->setUsername($username);
         $user->setEmail($email);
         $user->setRoles([$isAdmin ? 'ROLE_ADMIN' : 'ROLE_USER']);
@@ -194,26 +193,30 @@ class AddUserCommand extends Command
         if ($output->isVerbose()) {
             $this->io->comment(sprintf('New user database id: %d / Elapsed time: %.2f ms / Consumed memory: %.2f MB', $user->getId(), $event->getDuration(), $event->getMemory() / (1024 ** 2)));
         }
+
+        return null;
     }
 
-    private function validateUserData($username, $plainPassword, $email, $fullName)
+    private function validateUserData(string $username, string $plainPassword, string $email, string $displayName): void
     {
         // first check if a user with the same username already exists.
         $existingUser = $this->users->findOneBy(['username' => $username]);
 
-        if (null !== $existingUser) {
+        if ($existingUser !== null) {
             throw new RuntimeException(sprintf('There is already a user registered with the "%s" username.', $username));
         }
+
+        // @todo Validation must be moved to a separate UserValidator
 
         // validate password and email if is not this input means interactive.
         $this->validator->validatePassword($plainPassword);
         $this->validator->validateEmail($email);
-        $this->validator->validateFullName($fullName);
+        $this->validator->validateDisplayName($displayName);
 
         // check if a user with the same email already exists.
         $existingEmail = $this->users->findOneBy(['email' => $email]);
 
-        if (null !== $existingEmail) {
+        if ($existingEmail !== null) {
             throw new RuntimeException(sprintf('There is already a user registered with the "%s" email.', $email));
         }
     }
