@@ -48,7 +48,7 @@ class ListUsersCommand extends Command
     /**
      * {@inheritdoc}
      */
-    protected function configure()
+    protected function configure(): void
     {
         $this
             ->setDescription('Lists all the existing users')
@@ -73,8 +73,7 @@ HELP
             // commands can optionally define arguments and/or options (mandatory and optional)
             // see https://symfony.com/doc/current/components/console/console_arguments.html
             ->addOption('max-results', null, InputOption::VALUE_OPTIONAL, 'Limits the number of users listed', 50)
-            ->addOption('send-to', null, InputOption::VALUE_OPTIONAL, 'If set, the result is sent to the given email address')
-        ;
+            ->addOption('send-to', null, InputOption::VALUE_OPTIONAL, 'If set, the result is sent to the given email address');
     }
 
     /**
@@ -91,7 +90,7 @@ HELP
         $usersAsPlainArrays = array_map(function (User $user) {
             return [
                 $user->getId(),
-                $user->getFullName(),
+                $user->getDisplayName(),
                 $user->getUsername(),
                 $user->getEmail(),
                 implode(', ', $user->getRoles()),
@@ -114,9 +113,12 @@ HELP
         $usersAsATable = $bufferedOutput->fetch();
         $output->write($usersAsATable);
 
-        if (null !== $email = $input->getOption('send-to')) {
+        $email = $input->getOption('send-to');
+        if ($email !== null) {
             $this->sendReport($usersAsATable, $email);
         }
+
+        return null;
     }
 
     /**
@@ -129,8 +131,7 @@ HELP
             ->setSubject(sprintf('app:list-users report (%s)', date('Y-m-d H:i:s')))
             ->setFrom($this->emailSender)
             ->setTo($recipient)
-            ->setBody($contents, 'text/plain')
-        ;
+            ->setBody($contents, 'text/plain');
 
         $this->mailer->send($message);
     }
