@@ -5,53 +5,90 @@ declare(strict_types=1);
 namespace Bolt\Configuration;
 
 use Bolt\Common\Arr;
+use Bolt\Configuration\Parser\BaseParser;
 use Bolt\Configuration\Parser\ContentTypesParser;
 use Bolt\Configuration\Parser\GeneralParser;
 use Bolt\Configuration\Parser\TaxonomyParser;
-use Symfony\Component\Config\FileLocator;
+use Psr\SimpleCache\CacheInterface;
+use Symfony\Component\Stopwatch\Stopwatch;
 use Tightenco\Collect\Support\Collection;
-use Webmozart\PathUtil\Path;
 
 class Config
 {
-    /** @var array */
+    /** @var Collection */
     protected $data;
-
-    /** @var FileLocator */
-    private $fileLocator;
 
     /** @var PathResolver */
     private $pathResolver;
 
-    public function __construct()
+    /** @var Stopwatch */
+    private $stopwatch;
+
+    /** @var CacheInterface */
+    private $cache;
+
+    /** @var string */
+    private $projectDir;
+
+    public function __construct(Stopwatch $stopwatch, string $projectDir, CacheInterface $cache)
     {
-        $this->pathResolver = new PathResolver(dirname(dirname(__DIR__)), []);
+        $this->stopwatch = $stopwatch;
+        $this->cache = $cache;
+        $this->projectDir = $projectDir;
+        $this->data = $this->getConfig();
 
-        $configDirectories = [dirname(dirname(__DIR__)) . '/config/bolt'];
-        $this->fileLocator = new FileLocator($configDirectories);
+        // @todo PathResolver shouldn't be part of Config. Refactor to separate class
+        $this->pathResolver = new PathResolver($projectDir, []);
+    }
 
-        // $this->cacheFile = $this->app['filesystem']->getFile('cache://config-cache.json');
+    private function getConfig(): Collection
+    {
+        $this->stopwatch->start('bolt.parseconfig');
 
-        $data = null;
-
-        // $data = $this->loadCache();
-        if ($data === null) {
-            $data = $this->parseConfig();
-
-            // If we have to reload the config, we will also want to make sure
-            // the DB integrity is checked.
-            // $this->app['schema.timer']->setCheckRequired();
+        if ($this->validCache()) {
+            $data = $this->getCache();
+        } else {
+            [$data, $timestamps] = $this->parseConfig();
+            $this->setCache($data, $timestamps);
         }
 
-        $this->data = $data;
+        $this->stopwatch->stop('bolt.parseconfig');
+
+        return $data;
+    }
+
+    private function validCache(): bool
+    {
+        if (! $this->cache->has('config_cache') || ! $this->cache->has('config_timestamps')) {
+            return false;
+        }
+
+        $timestamps = $this->cache->get('config_timestamps');
+
+        foreach ($timestamps as $filename => $timestamp) {
+            if (file_exists($filename) === false || filemtime($filename) > $timestamp) {
+                return false;
+            }
+        }
+
+        return true;
+    }
+
+    private function getCache(): Collection
+    {
+        return $this->cache->get('config_cache');
+    }
+
+    private function setCache(Collection $data, array $timestamps): void
+    {
+        $this->cache->set('config_cache', $data);
+        $this->cache->set('config_timestamps', $timestamps);
     }
 
     /**
      * Load the configuration from the various YML files.
-     *
-     * @return array
      */
-    public function parseConfig()
+    private function parseConfig(): array
     {
         $general = new GeneralParser();
 
@@ -59,20 +96,39 @@ class Config
             'general' => $general->parse(),
         ]);
 
-        $this->data = $config;
-
         $taxonomy = new TaxonomyParser();
         $config['taxonomies'] = $taxonomy->parse();
 
-        $contenttypes = new ContentTypesParser($this->get('general')['accept_file_types']);
-        $config['contenttypes'] = $contenttypes->parse();
+        $contentTypes = new ContentTypesParser($config->get('general')['accept_file_types']);
+        $config['contenttypes'] = $contentTypes->parse();
 
+        // @todo Add these config files if needed, or refactor them out otherwise
         //'menu' => $this->parseConfigYaml('menu.yml'),
         //'routing' => $this->parseConfigYaml('routing.yml'),
         //'permissions' => $this->parseConfigYaml('permissions.yml'),
         //'extensions' => $this->parseConfigYaml('extensions.yml'),
 
-        return $config;
+        $timestamps = $this->getConfigFilesTimestamps($general, $taxonomy, $contentTypes);
+
+        return [$config, $timestamps];
+    }
+
+    private function getConfigFilesTimestamps(BaseParser ...$configs): array
+    {
+        $timestamps = [];
+
+        foreach ($configs as $config) {
+            foreach ($config->getFilenames() as $file) {
+                $timestamps[$file] = filemtime($file);
+            }
+        }
+
+        $envFilename = $this->projectDir . '/.env';
+        if (file_exists($envFilename)) {
+            $timestamps[$envFilename] = filemtime($envFilename);
+        }
+
+        return $timestamps;
     }
 
     /**
@@ -81,41 +137,25 @@ class Config
      * For example:
      * $var = $config->get('general/wysiwyg/ck/contentsCss');
      *
-     * @param string            $path
      * @param string|array|bool $default
-     *
-     * @return mixed
      */
     public function get(string $path, $default = null)
     {
         return Arr::get($this->data, $path, $default);
     }
 
-    /**
-     * @param string $path
-     * @param bool   $absolute
-     * @param mixed  $additional
-     *
-     * @return string
-     */
     public function getPath(string $path, bool $absolute = true, $additional = null): string
     {
         return $this->pathResolver->resolve($path, $absolute, $additional);
     }
 
-    /**
-     * @return Collection
-     */
     public function getPaths(): Collection
     {
         return $this->pathResolver->resolveAll();
     }
 
-    /**
-     * @return Collection
-     */
     public function getMediaTypes(): Collection
     {
-        return collect(['png', 'jpg', 'jpeg', 'gif', 'svg', 'pdf', 'mp3', 'tiff']);
+        return collect($this->get('general/accept_media_types'));
     }
 }
