@@ -5,26 +5,33 @@ declare(strict_types=1);
 namespace Bolt\Configuration\Parser;
 
 use Bolt\Common\Arr;
-use Bolt\Helpers\Str;
-use Cocur\Slugify\Slugify;
+use Bolt\Enum\Statuses;
+use Bolt\Utils\Str;
 use Exception;
 use Tightenco\Collect\Support\Collection;
 
 class ContentTypesParser extends BaseParser
 {
-    private $exceptions;
+    /**
+     * @var string[]
+     */
+    private $acceptFileTypes = [];
+
+    public function __construct(array $acceptFileTypes)
+    {
+        $this->acceptFileTypes = $acceptFileTypes;
+        parent::__construct();
+    }
 
     /**
      * Read and parse the contenttypes.yml configuration file.
      *
      * @throws Exception
-     *
-     * @return Collection
      */
     public function parse(): Collection
     {
         $contentTypes = new Collection();
-        $tempContentTypes = $this->parseConfigYaml('contenttypes.yml');
+        $tempContentTypes = $this->parseConfigYaml('contenttypes.yaml');
         foreach ($tempContentTypes as $key => $contentType) {
             $contentType = $this->parseContentType($key, $contentType);
             $contentTypes[$key] = $contentType;
@@ -34,123 +41,131 @@ class ContentTypesParser extends BaseParser
     }
 
     /**
-     * Parse a single Contenttype configuration array.
+     * Parse a single Content Type configuration array.
      *
      * @param string $key
      * @param array  $contentType
      *
      * @throws \Exception
-     *
-     * @return array
      */
-    protected function parseContentType($key, $contentType)
+    protected function parseContentType($key, $contentType): array
     {
         // If the slug isn't set, and the 'key' isn't numeric, use that as the slug.
-        if (!isset($contentType['slug']) && !is_numeric($key)) {
-            $contentType['slug'] = Slugify::create()->slugify($key);
+        if (! isset($contentType['slug']) && ! is_numeric($key)) {
+            $contentType['slug'] = Str::slug($key);
         }
 
         // If neither 'name' nor 'slug' is set, we need to warn the user. Same goes for when
         // neither 'singular_name' nor 'singular_slug' is set.
-        if (!isset($contentType['name']) && !isset($contentType['slug'])) {
-            $error = sprintf("In contenttype <code>%s</code>, neither 'name' nor 'slug' is set. Please edit <code>contenttypes.yml</code>, and correct this.", $key);
+        if (! isset($contentType['name']) && ! isset($contentType['slug'])) {
+            $error = sprintf("In content type <code>%s</code>, neither 'name' nor 'slug' is set. Please edit <code>contenttypes.yml</code>, and correct this.", $key);
             throw new Exception($error);
         }
-        if (!isset($contentType['singular_name']) && !isset($contentType['singular_slug'])) {
-            $error = sprintf("In contenttype <code>%s</code>, neither 'singular_name' nor 'singular_slug' is set. Please edit <code>contenttypes.yml</code>, and correct this.", $key);
-            throw new Exception($error);
-        }
-
-        // Contenttypes without fields make no sense.
-        if (!isset($contentType['fields'])) {
-            $error = sprintf("In contenttype <code>%s</code>, no 'fields' are set. Please edit <code>contenttypes.yml</code>, and correct this.", $key);
+        if (! isset($contentType['singular_name']) && ! isset($contentType['singular_slug'])) {
+            $error = sprintf("In content type <code>%s</code>, neither 'singular_name' nor 'singular_slug' is set. Please edit <code>contenttypes.yml</code>, and correct this.", $key);
             throw new Exception($error);
         }
 
-        if (!isset($contentType['slug'])) {
-            $contentType['slug'] = Slugify::create()->slugify($contentType['name']);
+        // Content types without fields make no sense.
+        if (! isset($contentType['fields'])) {
+            $error = sprintf("In content type <code>%s</code>, no 'fields' are set. Please edit <code>contenttypes.yml</code>, and correct this.", $key);
+            throw new Exception($error);
         }
-        if (!isset($contentType['name'])) {
+
+        if (! isset($contentType['slug'])) {
+            $contentType['slug'] = Str::slug($contentType['name']);
+        }
+        if (! isset($contentType['name'])) {
             $contentType['name'] = ucwords(preg_replace('/[^a-z0-9]/i', ' ', $contentType['slug']));
         }
-        if (!isset($contentType['singular_slug'])) {
-            $contentType['singular_slug'] = Slugify::create()->slugify($contentType['singular_name']);
+        if (! isset($contentType['singular_slug'])) {
+            $contentType['singular_slug'] = Str::slug($contentType['singular_name']);
         }
-        if (!isset($contentType['singular_name'])) {
+        if (! isset($contentType['singular_name'])) {
             $contentType['singular_name'] = ucwords(preg_replace('/[^a-z0-9]/i', ' ', $contentType['singular_slug']));
         }
-        if (!isset($contentType['show_on_dashboard'])) {
+        if (! isset($contentType['show_on_dashboard'])) {
             $contentType['show_on_dashboard'] = true;
         }
-        if (!isset($contentType['show_in_menu'])) {
+        if (! isset($contentType['show_in_menu'])) {
             $contentType['show_in_menu'] = true;
         }
-        if (!isset($contentType['sort'])) {
+        if (! isset($contentType['sort'])) {
             $contentType['sort'] = false;
         }
-        if (!isset($contentType['default_status'])) {
-            $contentType['default_status'] = 'published';
+        if (! isset($contentType['default_status'])) {
+            $contentType['default_status'] = Statuses::PUBLISHED;
         }
-        if (!isset($contentType['viewless'])) {
+        if (! isset($contentType['viewless'])) {
             $contentType['viewless'] = false;
         }
-        if (!isset($contentType['icon_one'])) {
+        if (! isset($contentType['icon_one'])) {
             $contentType['icon_one'] = 'fa-file';
         } else {
             $contentType['icon_one'] = str_replace('fa:', 'fa-', $contentType['icon_one']);
         }
-        if (!isset($contentType['icon_many'])) {
+        if (! isset($contentType['icon_many'])) {
             $contentType['icon_many'] = 'fa-copy';
         } else {
             $contentType['icon_many'] = str_replace('fa:', 'fa-', $contentType['icon_many']);
         }
 
-        // Allow explicit setting of a Contenttype's table name suffix. We default
-        // to slug if not present as it has been this way since Bolt v1.2.1
-        if (!isset($contentType['tablename'])) {
-            $contentType['tablename'] = Slugify::create()->slugify($contentType['slug'], '_');
+        // Allow explicit setting of a Content Type's table name suffix. We default to slug if not present.
+        if (isset($contentType['tablename'])) {
+            $contentType['tablename'] = Str::slug($contentType['tablename'], '_');
         } else {
-            $contentType['tablename'] = Slugify::create()->slugify($contentType['tablename'], '_');
+            $contentType['tablename'] = Str::slug($contentType['slug'], '_');
         }
-        if (!isset($contentType['allow_numeric_slugs'])) {
+
+        if (! isset($contentType['allow_numeric_slugs'])) {
             $contentType['allow_numeric_slugs'] = false;
         }
-        if (!isset($contentType['singleton'])) {
+        if (! isset($contentType['singleton'])) {
             $contentType['singleton'] = false;
         }
 
-        list($fields, $groups) = $this->parseFieldsAndGroups($contentType['fields']);
+        if (! isset($contentType['locales'])) {
+            $contentType['locales'] = [];
+        } elseif (is_string($contentType['locales'])) {
+            $contentType['locales'] = (array) $contentType['locales'];
+        }
+
+        [$fields, $groups] = $this->parseFieldsAndGroups($contentType['fields']);
         $contentType['fields'] = $fields;
         $contentType['groups'] = $groups;
 
         // Make sure taxonomy is an array.
         if (isset($contentType['taxonomy'])) {
             $contentType['taxonomy'] = (array) $contentType['taxonomy'];
+        } else {
+            $contentType['taxonomy'] = [];
         }
 
         // when adding relations, make sure they're added by their slug. Not their 'name' or 'singular name'.
-        if (!empty($contentType['relations']) && is_array($contentType['relations'])) {
+        if (! empty($contentType['relations']) && is_array($contentType['relations'])) {
             foreach (array_keys($contentType['relations']) as $relkey) {
-                if ($relkey !== Slugify::create()->slugify($relkey)) {
-                    $contentType['relations'][Slugify::create()->slugify($relkey)] = $contentType['relations'][$relkey];
+                if ($relkey !== Str::slug($relkey)) {
+                    $contentType['relations'][Str::slug($relkey)] = $contentType['relations'][$relkey];
                     unset($contentType['relations'][$relkey]);
                 }
             }
+        } else {
+            $contentType['relations'] = [];
+        }
+
+        if (! empty($contentType['relations']) || ! empty($contentType['taxonomy'])) {
+            $contentType['groups'][] = 'Relations';
         }
 
         return $contentType;
     }
 
     /**
-     * Parse a Contenttype's field and determine the grouping.
-     *
-     * @param array $fields
+     * Parse a Content Type's field and determine the grouping.
      *
      * @throws Exception
-     *
-     * @return array
      */
-    protected function parseFieldsAndGroups(array $fields)
+    protected function parseFieldsAndGroups(array $fields): array
     {
         $currentGroup = 'ungrouped';
         $groups = [];
@@ -159,7 +174,7 @@ class ContentTypesParser extends BaseParser
         foreach ($fields as $key => $field) {
             unset($fields[$key]);
             $key = str_replace('-', '_', mb_strtolower(Str::makeSafe($key, true)));
-            if (!isset($field['type']) || empty($field['type'])) {
+            if (! isset($field['type']) || empty($field['type'])) {
                 $error = sprintf('Field "%s" has no "type" set.', $key);
 
                 throw new \Exception($error);
@@ -168,7 +183,7 @@ class ContentTypesParser extends BaseParser
             // If field is a "file" type, make sure the 'extensions' are set, and it's an array.
             if ($field['type'] === 'file' || $field['type'] === 'filelist') {
                 if (empty($field['extensions'])) {
-                    $field['extensions'] = $this->accept_file_types;
+                    $field['extensions'] = $this->acceptFileTypes;
                 }
 
                 $field['extensions'] = (array) $field['extensions'];
@@ -178,7 +193,7 @@ class ContentTypesParser extends BaseParser
             if ($field['type'] === 'image' || $field['type'] === 'imagelist') {
                 if (empty($field['extensions'])) {
                     $field['extensions'] = collect(['gif', 'jpg', 'jpeg', 'png', 'svg'])
-                        ->intersect($this->accept_file_types);
+                        ->intersect($this->acceptFileTypes);
                 }
 
                 $field['extensions'] = (array) $field['extensions'];
@@ -190,7 +205,7 @@ class ContentTypesParser extends BaseParser
                 $field['values'] = array_combine($field['values'], $field['values']);
             }
 
-            if (!empty($field['group'])) {
+            if (! empty($field['group'])) {
                 $hasGroups = true;
             }
 
@@ -202,6 +217,7 @@ class ContentTypesParser extends BaseParser
                     'group' => $currentGroup,
                     'label' => '',
                     'variant' => '',
+                    'localize' => false,
                 ],
                 $field
             );
@@ -215,7 +231,7 @@ class ContentTypesParser extends BaseParser
 
             // Repeating fields checks
             if ($field['type'] === 'repeater') {
-                $fields[$key] = $this->parseFieldRepeaters($fields, $key);
+                $fields[$key] = $this->parseFieldRepeaters($fields[$key]);
                 if ($fields[$key] === null) {
                     unset($fields[$key]);
                 }
@@ -232,23 +248,17 @@ class ContentTypesParser extends BaseParser
 
     /**
      * Basic validation of repeater fields.
-     *
-     * @param array  $fields
-     * @param string $key
-     *
-     * @return array
      */
-    private function parseFieldRepeaters(array $fields, $key)
+    private function parseFieldRepeaters(array $repeater): array
     {
         $blacklist = ['repeater', 'slug', 'templatefield'];
-        $repeater = $fields[$key];
 
-        if (!isset($repeater['fields']) || !is_array($repeater['fields'])) {
-            return;
+        if (! isset($repeater['fields']) || ! is_array($repeater['fields'])) {
+            return [];
         }
 
         foreach ($repeater['fields'] as $repeaterKey => $repeaterField) {
-            if (!isset($repeaterField['type']) || in_array($repeaterField['type'], $blacklist, true)) {
+            if (! isset($repeaterField['type']) || in_array($repeaterField['type'], $blacklist, true)) {
                 unset($repeater['fields'][$repeaterKey]);
             }
         }
