@@ -9,10 +9,13 @@ use Bolt\Entity\Media;
 use Bolt\Repository\MediaRepository;
 use Carbon\Carbon;
 use Faker\Factory;
+use Faker\Generator;
+use PHPExif\Exif;
 use PHPExif\Reader\Reader;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\Finder\SplFileInfo;
-use Webmozart\PathUtil\Path;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorage;
+use Tightenco\Collect\Support\Collection;
 
 class MediaFactory
 {
@@ -25,12 +28,17 @@ class MediaFactory
     /** @var ContainerInterface */
     private $container;
 
+    /** @var Reader */
+    private $exif;
+
+    /** @var Generator */
+    private $faker;
+
+    /** @var Collection */
+    private $mediaTypes;
+
     /**
      * MediaFactory constructor.
-     *
-     * @param Config             $config
-     * @param MediaRepository    $mediaRepository
-     * @param ContainerInterface $container
      */
     public function __construct(Config $config, MediaRepository $mediaRepository, ContainerInterface $container)
     {
@@ -40,27 +48,30 @@ class MediaFactory
 
         $this->exif = Reader::factory(Reader::TYPE_NATIVE);
         $this->faker = Factory::create();
-        $this->mediatypes = $config->getMediaTypes();
+        $this->mediaTypes = $config->getMediaTypes();
     }
 
     /**
-     * @param SplFileInfo $file
-     * @param string      $area
-     *
-     * @return Media
+     * @throws \Exception
      */
     public function createOrUpdateMedia(SplFileInfo $file, string $area): Media
     {
         $media = $this->mediaRepository->findOneBy([
             'area' => $area,
             'path' => $file->getRelativePath(),
-            'filename' => $file->getFilename(), ]);
+            'filename' => $file->getFilename(),
+        ]);
 
-        if (!$media) {
+        if (! $media) {
             $media = new Media();
             $media->setFilename($file->getFilename())
                 ->setPath($file->getRelativePath())
                 ->setArea($area);
+        }
+
+        if (! $this->mediaTypes->contains($file->getExtension())) {
+            // @todo We're throwing a generic Exception here. Needs to be handled better.
+            throw new \Exception('Not a valid media type.');
         }
 
         $media->setType($file->getExtension())
@@ -77,16 +88,12 @@ class MediaFactory
         return $media;
     }
 
-    /**
-     * @param Media $media
-     * @param $file
-     */
-    private function updateImageData(Media $media, $file)
+    private function updateImageData(Media $media, SplFileInfo $file): void
     {
-        /** @var Exif $exif */
+        /** @var Exif|bool $exif */
         $exif = $this->exif->read($file->getRealPath());
 
-        if ($exif) {
+        if ($exif instanceof Exif) {
             $media->setWidth($exif->getWidth())
                 ->setHeight($exif->getHeight());
 
@@ -103,51 +110,43 @@ class MediaFactory
         }
     }
 
-    /**
-     * @param $media
-     *
-     * @return bool
-     */
-    private function isImage($media)
+    private function isImage(Media $media): bool
     {
         return in_array($media->getType(), ['gif', 'png', 'jpg', 'svg'], true);
     }
 
     /**
-     * @return object|string|void
+     * @todo Refactor this out!
+     *
+     * @return object|string|null
      */
     protected function getUser()
     {
-        if (!$this->container->has('security.token_storage')) {
+        if (! $this->container->has('security.token_storage')) {
             throw new \LogicException('The SecurityBundle is not registered in your application. Try running "composer require symfony/security-bundle".');
         }
 
-        if (null === $token = $this->container->get('security.token_storage')->getToken()) {
-            return;
+        /** @var TokenStorage $tokenStorage */
+        $tokenStorage = $this->container->get('security.token_storage');
+
+        $token = $tokenStorage->getToken();
+        if ($token === null) {
+            return null;
         }
 
-        if (!\is_object($user = $token->getUser())) {
-            // e.g. anonymous authentication
-            return;
+        if (is_object($token->getUser())) {
+            return $token->getUser();
         }
 
-        return $user;
+        // e.g. anonymous authentication
+        return null;
     }
 
-    /**
-     * @param $area
-     * @param $path
-     * @param $filename
-     *
-     * @return Media
-     */
     public function createFromFilename($area, $path, $filename): Media
     {
         $target = $this->config->getPath($area, true, [$path, $filename]);
         $file = new SplFileInfo($target, $path, $filename);
 
-        $media = $this->createOrUpdateMedia($file, $area);
-
-        return $media;
+        return $this->createOrUpdateMedia($file, $area);
     }
 }
