@@ -10,6 +10,7 @@ use Bolt\Repository\ContentRepository;
 use Knp\Menu\FactoryInterface;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Stopwatch\Stopwatch;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 class MenuBuilder
 {
@@ -23,178 +24,239 @@ class MenuBuilder
     private $stopwatch;
 
     /** @var ContentRepository */
-    private $content;
+    private $contentRepository;
 
     /** @var UrlGeneratorInterface */
     private $urlGenerator;
 
+    /** @var TranslatorInterface */
+    private $translator;
+
     /**
      * MenuBuilder constructor.
-     *
-     * @param FactoryInterface      $factory
-     * @param Config                $config
-     * @param Stopwatch             $stopwatch
-     * @param ContentRepository     $content
-     * @param UrlGeneratorInterface $urlGenerator
      */
-    public function __construct(FactoryInterface $factory, Config $config, Stopwatch $stopwatch, ContentRepository $content, UrlGeneratorInterface $urlGenerator)
+    public function __construct(FactoryInterface $factory, Config $config, Stopwatch $stopwatch, ContentRepository $contentRepository, UrlGeneratorInterface $urlGenerator, TranslatorInterface $translator)
     {
         $this->config = $config;
         $this->factory = $factory;
         $this->stopwatch = $stopwatch;
-        $this->content = $content;
+        $this->contentRepository = $contentRepository;
         $this->urlGenerator = $urlGenerator;
+        $this->translator = $translator;
     }
 
     public function createSidebarMenu()
     {
         $this->stopwatch->start('bolt.sidebar');
 
+        $t = $this->translator;
+
         $menu = $this->factory->createItem('root');
 
-        $menu->addChild('Dashboard', ['uri' => 'homepage', 'extras' => [
-            'name' => 'Dashboard',
-            'icon' => 'fa-tachometer-alt',
-            'link' => '/bolt/',
-        ]]);
+        $menu->addChild('Dashboard', [
+            'uri' => $this->urlGenerator->generate('bolt_dashboard'),
+            'extras' => [
+                'name' => $t->trans('caption.dashboard'),
+                'icon' => 'fa-tachometer-alt',
+            ],
+        ]);
 
-        $menu->addChild('Content', ['uri' => 'content', 'extras' => [
-            'name' => 'Content',
+        $menu->addChild('Content', ['extras' => [
+            'name' => $t->trans('caption.content'),
             'type' => 'separator',
             'icon' => 'fa-file',
         ]]);
 
-        $contenttypes = $this->config->get('contenttypes');
+        $contentTypes = $this->config->get('contenttypes');
 
-        foreach ($contenttypes as $contenttype) {
-            $menu->addChild($contenttype['slug'], ['uri' => 'homepage', 'extras' => [
-                'name' => $contenttype['name'],
-                'singular_name' => $contenttype['singular_name'],
-                'slug' => $contenttype['slug'],
-                'singular_slug' => $contenttype['singular_slug'],
-                'icon' => $contenttype['icon_many'],
-                'link' => $this->urlGenerator->generate('bolt_contentlisting', ['contenttype' => $contenttype['slug']]),
-                'link_new' => $this->urlGenerator->generate('bolt_edit_record', ['id' => $contenttype['slug']]),
-                'contenttype' => $contenttype['slug'],
-                'singleton' => $contenttype['singleton'],
-                'active' => $contenttype['slug'] === 'pages' ? true : false,
-                'submenu' => $this->getLatestRecords($contenttype['slug']),
-            ]]);
+        foreach ($contentTypes as $contentType) {
+            $menu->addChild($contentType['slug'], [
+                'uri' => $this->urlGenerator->generate('bolt_content_overview', ['contentType' => $contentType['slug']]),
+                'extras' => [
+                    'name' => $contentType['name'],
+                    'singular_name' => $contentType['singular_name'],
+                    'slug' => $contentType['slug'],
+                    'singular_slug' => $contentType['singular_slug'],
+                    'icon' => $contentType['icon_many'],
+                    'link_new' => $this->urlGenerator->generate('bolt_content_edit', ['id' => $contentType['slug']]),
+                    'content_type' => $contentType['slug'],
+                    'singleton' => $contentType['singleton'],
+                    'active' => $contentType['slug'] === 'pages' ? true : false,
+                    'submenu' => $this->getLatestRecords($contentType['slug']),
+                ],
+            ]);
         }
 
-        $menu->addChild('Settings', ['uri' => 'settings', 'extras' => [
-            'name' => 'Settings',
+        $menu->addChild('Settings', ['extras' => [
+            'name' => $t->trans('caption.settings'),
             'type' => 'separator',
             'icon' => 'fa-wrench',
         ]]);
 
         // Configuration submenu
 
-        $menu->addChild('Configuration', ['uri' => 'configuration', 'extras' => [
-            'name' => 'Configuration',
-            'icon' => 'fa-flag',
-            'link' => '/bolt/finder/config',
+        $menu->addChild('Configuration', ['extras' => [
+            'name' => $t->trans('caption.configuration'),
+            'icon' => 'fa-sliders-h',
         ]]);
 
-        $menu['Configuration']->addChild('Users &amp; Permissions', ['uri' => 'users', 'extras' => [
-            'name' => 'Users &amp; Permissions',
-            'icon' => 'fa-group',
-            'link' => '/bolt/finder/config',
+        $menu->getChild('Configuration')->addChild('Users &amp; Permissions', [
+            'uri' => $this->urlGenerator->generate('bolt_users'),
+            'extras' => [
+                'name' => $t->trans('caption.users_permissions'),
+                'icon' => 'fa-users',
+            ],
+        ]);
+
+        $menu->getChild('Configuration')->addChild('Main configuration', [
+            'uri' => $this->urlGenerator->generate('bolt_file_edit', [
+                'area' => 'config',
+                'file' => '/bolt/config.yaml',
+            ]),
+            'extras' => [
+                'name' => $t->trans('caption.main_configuration'),
+                'icon' => 'fa-cog',
+            ],
+        ]);
+
+        $menu->getChild('Configuration')->addChild('ContentTypes', [
+            'uri' => $this->urlGenerator->generate('bolt_file_edit', [
+                'area' => 'config',
+                'file' => '/bolt/contenttypes.yaml',
+            ]),
+            'extras' => [
+                'name' => $t->trans('caption.contenttypes'),
+                'icon' => 'fa-object-group',
+            ],
+        ]);
+
+        $menu->getChild('Configuration')->addChild('Taxonomies', [
+            'uri' => $this->urlGenerator->generate('bolt_file_edit', [
+                'area' => 'config',
+                'file' => '/bolt/taxonomy.yaml',
+            ]),
+            'extras' => [
+                'name' => $t->trans('caption.taxonomies'),
+                'icon' => 'fa-tags',
+            ],
+        ]);
+
+        $menu->getChild('Configuration')->addChild('Menu set up', [
+            'uri' => $this->urlGenerator->generate('bolt_file_edit', [
+                'area' => 'config',
+                'file' => '/bolt/menu.yaml',
+            ]),
+            'extras' => [
+                'name' => $t->trans('caption.menu_setup'),
+                'type' => 'separator',
+                'icon' => 'fa-list',
+            ],
+        ]);
+
+        $menu->getChild('Configuration')->addChild('Routing set up', [
+            'uri' => $this->urlGenerator->generate('bolt_file_edit', [
+                'area' => 'config',
+                'file' => '/routes.yaml',
+            ]),
+            'extras' => [
+                'name' => $t->trans('caption.routing_setup'),
+                'icon' => 'fa-directions',
+            ],
+        ]);
+
+        $menu->getChild('Configuration')->addChild('All configuration files', [
+            'uri' => $this->urlGenerator->generate('bolt_filemanager', ['area' => 'config']),
+            'extras' => [
+                'name' => $t->trans('caption.all_configuration_files'),
+                'icon' => 'fa-cogs',
+            ],
+        ]);
+
+        // Maintenance submenu
+
+        $menu->addChild('Maintenance', ['extras' => [
+            'name' => $t->trans('caption.maintenance'),
+            'icon' => 'fa-tools',
         ]]);
 
-        $menu['Configuration']->addChild('Main configuration', ['uri' => 'config', 'extras' => [
-            'name' => 'Main configuration',
-            'icon' => 'fa-cog',
-            'link' => '/bolt/editfile/config?file=/bolt/config.yaml',
-        ]]);
+        $menu->getChild('Maintenance')->addChild('Bolt API', [
+            'uri' => $this->urlGenerator->generate('api_entrypoint'),
+            'extras' => [
+                'name' => $t->trans('caption.api'),
+                'icon' => 'fa-code',
+            ],
+        ]);
 
-        $menu['Configuration']->addChild('ContentTypes', ['uri' => 'contenttypes', 'extras' => [
-            'name' => 'ContentTypes',
-            'icon' => 'fa-paint-brush',
-            'link' => '/bolt/editfile/config?file=/bolt/contenttypes.yml',
-        ]]);
+        $menu->getChild('Maintenance')->addChild('Check database', [
+            'uri' => '',
+            'extras' => [
+                'name' => $t->trans('caption.check_database'),
+                'icon' => 'fa-database',
+            ],
+        ]);
 
-        $menu['Configuration']->addChild('Taxonomy', ['uri' => 'taxonomy', 'extras' => [
-            'name' => 'Taxonomy',
-            'icon' => 'fa-tags',
-            'link' => '/bolt/editfile/config?file=/bolt/taxonomy.yml',
-        ]]);
+        $menu->getChild('Maintenance')->addChild('Fixtures', [
+            'uri' => '',
+            'extras' => [
+                'name' => $t->trans('caption.fixtures_dummy_content'),
+                'icon' => 'fa-hat-wizard',
+            ],
+        ]);
 
-        $menu['Configuration']->addChild('Menu set up', ['uri' => 'menusetup', 'extras' => [
-            'name' => 'Menu set up',
-            'type' => 'separator',
-            'icon' => 'fa-list',
-            'link' => '/bolt/editfile/config?file=/bolt/menu.yml',
-        ]]);
+        $menu->getChild('Maintenance')->addChild('Clear the cache', [
+            'uri' => $this->urlGenerator->generate('bolt_clear_cache'),
+            'extras' => [
+                'name' => $t->trans('caption.clear_cache'),
+                'icon' => 'fa-eraser',
+            ],
+        ]);
 
-        $menu['Configuration']->addChild('Routing set up', ['uri' => 'routing', 'extras' => [
-            'name' => 'Routing set up',
-            'icon' => 'fa-random',
-            'link' => '/bolt/editfile/config?file=/bolt/routing.yml',
-        ]]);
+        $menu->getChild('Maintenance')->addChild('Installation checks', [
+            'uri' => '',
+            'extras' => [
+                'name' => $t->trans('caption.installation_checks'),
+                'icon' => 'fa-clipboard-check',
+            ],
+        ]);
 
-        $menu['Configuration']->addChild('Check database', ['uri' => 'database', 'extras' => [
-            'name' => 'Check database',
-            'type' => 'separator',
-            'icon' => 'fa-database',
-            'link' => '/bolt/finder/config',
-        ]]);
+        $menu->getChild('Maintenance')->addChild('Translations: Messages', [
+            'uri' => $this->urlGenerator->generate('translation_index'),
+            'extras' => [
+                'name' => $t->trans('caption.translations'),
+                'icon' => 'fa-language',
+            ],
+        ]);
 
-        $menu['Configuration']->addChild('Clear the cache', ['uri' => 'cache', 'extras' => [
-            'name' => 'Clear the cache',
-            'icon' => 'fa-eraser',
-            'link' => '/bolt/finder/config',
-        ]]);
-
-        $menu['Configuration']->addChild('Change Log', ['uri' => 'else', 'extras' => [
-            'name' => 'Change Log',
-            'icon' => 'fa-archive',
-            'link' => '/bolt/finder/config',
-        ]]);
-
-        $menu['Configuration']->addChild('System Log', ['uri' => 'else', 'extras' => [
-            'name' => 'System Log',
-            'icon' => 'fa-archive',
-            'link' => '/bolt/finder/config',
-        ]]);
-
-        $menu['Configuration']->addChild('Set-up checks', ['uri' => 'else', 'extras' => [
-            'name' => 'Set-up checks',
-            'icon' => 'fa-support',
-            'link' => '/bolt/finder/config',
-        ]]);
-
-        $menu['Configuration']->addChild('Translations: Messages', ['uri' => 'else', 'extras' => [
-            'name' => 'Translations: Messages',
-            'type' => 'separator',
-            'icon' => 'fa-flag',
-            'link' => '/bolt/finder/config',
-        ]]);
+        $menu->getChild('Maintenance')->addChild('Extensions', [
+            'uri' => '',
+            'extras' => [
+                'name' => $t->trans('caption.extensions'),
+                'icon' => 'fa-plug',
+            ],
+        ]);
 
         // File Management submenu
-        $menu->addChild('File Management', ['uri' => 'content-files', 'extras' => [
-            'name' => 'File Management',
-            'icon' => 'fa-flag',
-            'link' => '/bolt/finder/files',
+
+        $menu->addChild('File Management', ['extras' => [
+            'name' => $t->trans('caption.file_management'),
+            'icon' => 'fa-folder-open',
         ]]);
 
-        $menu['File Management']->addChild('Uploaded files', ['uri' => 'content-files', 'extras' => [
-            'name' => 'Uploaded files',
-            'icon' => 'fa-folder-open-o',
-            'link' => '/bolt/finder/files',
-        ]]);
+        $menu->getChild('File Management')->addChild('Uploaded files', [
+            'uri' => $this->urlGenerator->generate('bolt_filemanager', ['area' => 'files']),
+            'extras' => [
+                'name' => $t->trans('caption.uploaded_files'),
+                'icon' => 'fa-archive',
+            ],
+        ]);
 
-        $menu['File Management']->addChild('View/edit Templates', ['uri' => 'theme-files', 'extras' => [
-            'name' => 'View/edit Templates',
-            'icon' => 'fa-desktop',
-            'link' => '/bolt/finder/themes',
-        ]]);
-
-        $menu->addChild('Extensions', ['uri' => 'extensions', 'extras' => [
-            'name' => 'Extensions',
-            'icon' => 'fa-cubes',
-            'link' => '/bolt/extensions',
-        ]]);
+        $menu->getChild('File Management')->addChild('View/edit Templates', [
+            'uri' => $this->urlGenerator->generate('bolt_filemanager', ['area' => 'themes']),
+            'extras' => [
+                'name' => $t->trans('caption.view_edit_templates'),
+                'icon' => 'fa-scroll',
+            ],
+        ]);
 
         $this->stopwatch->stop('bolt.sidebar');
 
@@ -204,10 +266,10 @@ class MenuBuilder
     private function getLatestRecords($slug)
     {
         /** @var ContentType $ct */
-        $contenttype = ContentTypeFactory::get($slug, $this->config->get('contenttypes'));
+        $contentType = ContentType::factory($slug, $this->config->get('contenttypes'));
 
-        /** @var Content $records */
-        $records = $this->content->findLatest($contenttype, 5);
+        /** @var Content[] $records */
+        $records = $this->contentRepository->findLatest($contentType, 5);
 
         $result = [];
 
@@ -242,7 +304,7 @@ class MenuBuilder
                         'slug' => $submenuChild->getExtra('slug'),
                         'singular_slug' => $submenuChild->getExtra('singular_slug'),
                         'icon' => $submenuChild->getExtra('icon'),
-                        'link' => $submenuChild->getExtra('link'),
+                        'link' => $submenuChild->getUri(),
                         'link_new' => $submenuChild->getExtra('link_new'),
                         'contenttype' => $submenuChild->getExtra('contenttype'),
                         'singleton' => $submenuChild->getExtra('singleton'),
@@ -260,7 +322,7 @@ class MenuBuilder
                 'slug' => $child->getExtra('slug'),
                 'singular_slug' => $child->getExtra('singular_slug'),
                 'icon' => $child->getExtra('icon'),
-                'link' => $child->getExtra('link'),
+                'link' => $child->getUri(),
                 'link_new' => $child->getExtra('link_new'),
                 'contenttype' => $child->getExtra('contenttype'),
                 'singleton' => $child->getExtra('singleton'),
