@@ -30,16 +30,16 @@ final class News
     /**
      * News. Film at 11.
      *
-     * @param Request $request
-     *
-     * @throws \GuzzleHttp\Exception\GuzzleException
-     *
-     * @return JsonResponse
      * @Route("/news", name="bolt_news")
      */
-    public function dashboardNews(Request $request)
+    public function dashboardNews(Request $request): JsonResponse
     {
         $news = $this->getNews($request->getHost());
+
+        // @todo Make sure this works as intended
+        if ($this->config->get('news') !== false) {
+            return new JsonResponse([], 200);
+        }
 
         // One 'alert' and one 'info' max. Regular info-items can be disabled,
         // but Alerts can't.
@@ -51,21 +51,15 @@ final class News
             'disable' => false, // $this->getOption('general/backend/news/disable'),
         ];
 
-        $response = new JsonResponse($context, 200);
-
-        return $response;
+        return new JsonResponse($context, 200);
     }
 
     /**
      * Get the news from Bolt HQ (with caching).
      *
      * @param string $hostname
-     *
-     * @throws \GuzzleHttp\Exception\GuzzleException
-     *
-     * @return array
      */
-    private function getNews($hostname)
+    private function getNews($hostname): array
     {
 //        // Cached for two hours.
 //        $news = $this->app['cache']->fetch('dashboardnews');
@@ -76,23 +70,15 @@ final class News
 //        }
 
         // If not cached, get fresh news.
-        $news = $this->fetchNews($hostname);
-
-//        $this->app['cache']->save('dashboardnews', $news, 7200);
-
-        return $news;
+        return $this->fetchNews($hostname);
     }
 
     /**
      * Get the news from Bolt HQ.
      *
      * @param string $hostname
-     *
-     * @throws \GuzzleHttp\Exception\GuzzleException
-     *
-     * @return array
      */
-    private function fetchNews($hostname)
+    private function fetchNews($hostname): array
     {
 //        $source = $this->getOption('general/branding/news_source', 'https://news.bolt.cm/');
         $source = 'https://news.bolt.cm/';
@@ -102,18 +88,18 @@ final class News
 
         try {
             $client = new Client(['base_uri' => $source]);
-            $fetchedNewsData = $client->request('GET', '/', $options)->getBody();
+            $fetchedNewsData = $client->request('GET', '/', $options)->getBody()->getContents();
         } catch (RequestException $e) {
-            $this->app['logger.system']->error(
-                'Error occurred during newsfeed fetch',
-                ['event' => 'exception', 'exception' => $e]
-            );
+//            $this->app['logger.system']->error(
+//                'Error occurred during newsfeed fetch',
+//                ['event' => 'exception', 'exception' => $e]
+//            );
 
             return [
                 'error' => [
                     'type' => 'error',
                     'title' => 'Unable to fetch news!',
-                    'teaser' => "<p>Unable to connect to $source</p>",
+                    'teaser' => "<p>Unable to connect to ${source}</p>",
                 ],
             ];
         }
@@ -136,8 +122,8 @@ final class News
         // applies and the first alert we need to show
         foreach ($fetchedNewsItems as $item) {
             $type = isset($item->type) ? $item->type : 'information';
-            if (!isset($news[$type])
-                && (empty($item->target_version) || Bolt\Version::compare($item->target_version, '>'))
+            if (! isset($news[$type])
+                && (empty($item->target_version) || Version::compare($item->target_version, '>'))
             ) {
                 $news[$type] = $item;
             }
@@ -152,7 +138,7 @@ final class News
             'error' => [
                 'type' => 'error',
                 'title' => 'Unable to fetch news!',
-                'teaser' => "<p>Invalid JSON feed returned by $source</p>",
+                'teaser' => "<p>Invalid JSON feed returned by ${source}</p>",
             ],
         ];
     }
@@ -161,14 +147,12 @@ final class News
      * Get the guzzle options.
      *
      * @param string $hostname
-     *
-     * @return array
      */
-    private function fetchNewsOptions($hostname)
+    private function fetchNewsOptions($hostname): array
     {
 //        $driver = $this->app['db']->getDatabasePlatform()->getName();
 
-        $options = [
+        return [
             'query' => [
                 'v' => Version::VERSION,
                 'p' => PHP_VERSION,
@@ -178,16 +162,5 @@ final class News
             'connect_timeout' => 5,
             'timeout' => 10,
         ];
-
-//        if ($this->getOption('general/httpProxy')) {
-//            $options['proxy'] = sprintf(
-//                '%s:%s@%s',
-//                $this->getOption('general/httpProxy/user'),
-//                $this->getOption('general/httpProxy/password'),
-//                $this->getOption('general/httpProxy/host')
-//            );
-//        }
-
-        return $options;
     }
 }
