@@ -12,24 +12,25 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
+ * Class DashboardController.
+ *
  * @Security("has_role('ROLE_ADMIN')")
  */
 class DashboardController extends BaseController
 {
     /**
-     * @Route("/", name="bolt_dashboard")
-     * was: ("/{vueRouting}", requirements={"vueRouting"="^(?!api|_(profiler|wdt)).+"}, name="index")
+     * @Route("/", name="bolt_dashboard", methods={"GET"})
      *
-     * @param ContentRepository $content
-     *
-     * @return Response
+     * @throws \Twig_Error_Loader
+     * @throws \Twig_Error_Runtime
+     * @throws \Twig_Error_Syntax
      */
-    public function index(ContentRepository $content)
+    public function index(ContentRepository $content): Response
     {
         /** @var Content $records */
         $records = $content->findLatest();
 
-        return $this->renderTemplate('dashboard/dashboard.twig', [
+        return $this->renderTemplate('dashboard/dashboard.html.twig', [
             'records' => $records,
         ]);
     }
