@@ -5,13 +5,9 @@ declare(strict_types=1);
 namespace Bolt\Controller\Backend;
 
 use Bolt\Controller\BaseController;
-use Bolt\Form\ChangePasswordType;
-use Bolt\Form\UserType;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
-use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
 
 /**
  * Class UserController.
@@ -21,52 +17,15 @@ use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
 class UserController extends BaseController
 {
     /**
-     * @Route("/profile-edit", methods={"GET", "POST"}, name="bolt_profile_edit")
+     * @Route("/users", name="bolt_users")
      */
-    public function edit(Request $request): Response
+    public function users(): Response
     {
-        $user = $this->getUser();
-        $form = $this->createForm(UserType::class, $user);
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
-            $this->addFlash('success', 'user.updated_successfully');
+        $twigVars = [
+            'title' => 'Users & Permissions',
+            'subtitle' => 'To edit users and their permissions',
+        ];
 
-            return $this->redirectToRoute('bolt_profile_edit');
-        }
-
-        return $this->renderTemplate('users/edit.twig', [
-            'user' => $user,
-            'form' => $form->createView(),
-        ]);
-    }
-
-    /**
-     * @Route("/change-password", methods={"GET", "POST"}, name="bolt_change_password")
-     *
-     * @param Request                      $request
-     * @param UserPasswordEncoderInterface $encoder
-     *
-     * @throws \Twig_Error_Loader
-     * @throws \Twig_Error_Runtime
-     * @throws \Twig_Error_Syntax
-     *
-     * @return Response
-     */
-    public function changePassword(Request $request, UserPasswordEncoderInterface $encoder): Response
-    {
-        $user = $this->getUser();
-        $form = $this->createForm(ChangePasswordType::class);
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            $user->setPassword($encoder->encodePassword($user, $form->get('newPassword')->getData()));
-            $this->getDoctrine()->getManager()->flush();
-
-            return $this->redirectToRoute('security_logout');
-        }
-
-        return $this->renderTemplate('users/change_password.twig', [
-            'form' => $form->createView(),
-        ]);
+        return $this->renderTemplate('pages/placeholder.html.twig', $twigVars);
     }
 }
