@@ -4,38 +4,49 @@ declare(strict_types=1);
 
 namespace Bolt\Controller\Frontend;
 
+use Bolt\Configuration\Config;
 use Bolt\Controller\BaseController;
 use Bolt\Repository\ContentRepository;
 use Bolt\Repository\FieldRepository;
+use Bolt\TemplateChooser;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
 
 class DetailController extends BaseController
 {
+    public function __construct(Config $config, CsrfTokenManagerInterface $csrfTokenManager, TemplateChooser $templateChooser)
+    {
+        parent::__construct($config, $csrfTokenManager);
+
+        $this->templateChooser = $templateChooser;
+    }
+
     /**
      * @Route(
-     *     "/{contenttypeslug}/{slug}",
-     *     name="detail",
-     *     requirements={"contenttypeslug"="%bolt.requirement.contenttypes%"},
+     *     "/{contentTypeSlug}/{slugOrId}",
+     *     name="record",
+     *     requirements={"contentTypeSlug"="%bolt.requirement.contenttypes%"},
      *     methods={"GET"})
      *
-     * @param ContentRepository $contentRepository
-     * @param FieldRepository   $fieldRepository
-     * @param null              $slug
+     * @param string|int $slugOrId
      *
      * @throws \Twig_Error_Loader
      * @throws \Twig_Error_Runtime
      * @throws \Twig_Error_Syntax
-     *
-     * @return Response
      */
-    public function record(ContentRepository $contentRepository, FieldRepository $fieldRepository, string $contenttypeslug, string $slug): Response
+    public function record(ContentRepository $contentRepository, FieldRepository $fieldRepository, string $contentTypeSlug, $slugOrId): Response
     {
-        if (!is_numeric($slug)) {
-            $field = $fieldRepository->findOneBySlug($slug);
-            $record = $field->getContent();
+        if (is_numeric($slugOrId)) {
+            $record = $contentRepository->findOneBy(['id' => (int) $slugOrId]);
         } else {
-            $record = $contentRepository->findOneBy(['id' => $slug]);
+            /* @todo this should search only by slug or any other unique field */
+            $field = $fieldRepository->findOneBySlug($slugOrId);
+            if ($field === null) {
+                throw new NotFoundHttpException('Content does not exist.');
+            }
+            $record = $field->getContent();
         }
 
         $recordSlug = $record->getDefinition()['singular_slug'];
@@ -44,6 +55,9 @@ class DetailController extends BaseController
             'record' => $record,
             $recordSlug => $record,
         ];
+
+        dump($record);
+        dump($record->getFieldValues());
 
         $templates = $this->templateChooser->record($record);
 
