@@ -5,19 +5,21 @@ declare(strict_types=1);
 namespace Bolt\DataFixtures;
 
 use Bolt\Configuration\Config;
-use Bolt\Content\FieldFactory;
 use Bolt\Entity\Content;
+use Bolt\Entity\Field;
+use Bolt\Entity\User;
 use Doctrine\Bundle\FixturesBundle\Fixture;
 use Doctrine\Common\DataFixtures\DependentFixtureInterface;
 use Doctrine\Common\Persistence\ObjectManager;
 use Faker\Factory;
+use Tightenco\Collect\Support\Collection;
 
 class ContentFixtures extends Fixture implements DependentFixtureInterface
 {
     /** @var \Faker\Generator */
     private $faker;
 
-    /** @var Config */
+    /** @var Collection */
     private $config;
 
     private $lastTitle = null;
@@ -36,23 +38,25 @@ class ContentFixtures extends Fixture implements DependentFixtureInterface
         ];
     }
 
-    public function load(ObjectManager $manager)
+    public function load(ObjectManager $manager): void
     {
         $this->loadContent($manager);
 
         $manager->flush();
     }
 
-    private function loadContent(ObjectManager $manager)
+    private function loadContent(ObjectManager $manager): void
     {
         foreach ($this->config as $contentType) {
             $amount = $contentType['singleton'] ? 1 : 15;
 
             foreach (range(1, $amount) as $i) {
-                $author = $this->getReference(['jane_admin', 'tom_admin'][0 === $i ? 0 : random_int(0, 1)]);
+                $ref = $i === 0 ? 'admin' : ['admin', 'henkie', 'jane_admin', 'tom_admin'][random_int(0, 3)];
+                /** @var User $author */
+                $author = $this->getReference($ref);
 
                 $content = new Content();
-                $content->setContenttype($contentType['slug']);
+                $content->setContentType($contentType['slug']);
                 $content->setAuthor($author);
                 $content->setStatus($this->getRandomStatus());
                 $content->setCreatedAt($this->faker->dateTimeBetween('-1 year'));
@@ -62,12 +66,21 @@ class ContentFixtures extends Fixture implements DependentFixtureInterface
 
                 $sortorder = 1;
                 foreach ($contentType['fields'] as $name => $fieldType) {
-                    $field = FieldFactory::get($fieldType['type']);
-                    $field->setName($name);
-                    $field->setValue($this->getValuesforFieldType($name, $fieldType));
-                    $field->setSortorder($sortorder++ * 5);
+                    if ($fieldType['localize']) {
+                        $locales = $contentType['locales'];
+                    } else {
+                        $locales = [''];
+                    }
 
-                    $content->addField($field);
+                    foreach ($locales as $locale) {
+                        $field = Field::factory($fieldType, $name);
+                        $field->setName($name);
+                        $field->setValue($this->getValuesforFieldType($name, $fieldType));
+                        $field->setSortorder($sortorder++ * 5);
+                        $field->setLocale($locale);
+
+                        $content->addField($field);
+                    }
                 }
 
                 $manager->persist($content);
@@ -91,7 +104,10 @@ class ContentFixtures extends Fixture implements DependentFixtureInterface
                 $data = [$this->faker->paragraphs(3, true)];
                 break;
             case 'image':
-                $data = ['filename' => 'kitten.jpg', 'alt' => 'A cute kitten'];
+                $data = [
+                    'filename' => 'kitten.jpg',
+                    'alt' => 'A cute kitten',
+                ];
                 break;
             case 'slug':
                 $data = $this->lastTitle ?? [$this->faker->sentence(3, true)];
