@@ -7,41 +7,54 @@ namespace Bolt\Entity;
 use ApiPlatform\Core\Annotation\ApiFilter;
 use ApiPlatform\Core\Annotation\ApiResource;
 use ApiPlatform\Core\Bridge\Doctrine\Orm\Filter\SearchFilter;
-use Bolt\Configuration\Config;
 use Bolt\Content\ContentType;
-use Bolt\Content\ContentTypeFactory;
+use Bolt\Enum\Statuses;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
+use Doctrine\Common\Persistence\ObjectManagerAware;
 use Doctrine\ORM\Mapping as ORM;
-use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
+use Symfony\Component\Serializer\Annotation\Groups;
+use Symfony\Component\Serializer\Annotation\MaxDepth;
+use Tightenco\Collect\Support\Collection as LaravelCollection;
 
 /**
  * @ApiResource(
+ *     normalizationContext={"groups"={"get_content"}, "enable_max_depth"=true},
+ *     denormalizationContext={"groups"={"put"}},
  *     collectionOperations={"get"},
- *     itemOperations={"get"}
+ *     itemOperations={"get",
+ *         "put"={
+ *             "denormalization_context"={"groups"={"put"}},
+ *         }
+ *     }
  * )
  * @ApiFilter(SearchFilter::class)
  * @ORM\Entity(repositoryClass="Bolt\Repository\ContentRepository")
  * @ORM\Table(name="bolt_content")
  * @ORM\HasLifecycleCallbacks
  */
-class Content
+class Content implements ObjectManagerAware
 {
-    use ContentMagicTraits;
+    use ContentLocalizeTrait;
+    use ContentMagicTrait;
 
-    public const NUM_ITEMS = 8;
-
-    public const STATUSES = ['published', 'held', 'timed', 'draft'];
+    public const NUM_ITEMS = 8; // @todo This can't be a const
 
     /**
+     * @var int
+     *
      * @ORM\Id()
      * @ORM\GeneratedValue()
      * @ORM\Column(type="integer")
+     * @Groups("get_content")
      */
     private $id;
 
     /**
+     * @var string
+     *
      * @ORM\Column(type="string", length=191, name="contenttype")
+     * @Groups("get_content")
      */
     private $contentType;
 
@@ -50,36 +63,54 @@ class Content
      *
      * @ORM\ManyToOne(targetEntity="Bolt\Entity\User")
      * @ORM\JoinColumn(nullable=false)
+     * @Groups("put")
      */
     private $author;
 
     /**
+     * @var ?string
+     *
      * @ORM\Column(type="string", length=191)
+     * @Groups("put")
      */
-    private $status;
+    private $status = null;
 
     /**
-     * @ORM\Column(type="datetime")
+     * @var \DateTimeInterface
+     *
+     * @ORM\Column(type="datetime", nullable=false)
      */
     private $createdAt;
 
     /**
-     * @ORM\Column(type="datetime")
-     */
-    private $modifiedAt;
-
-    /**
+     * @var ?\DateTimeInterface
+     *
      * @ORM\Column(type="datetime", nullable=true)
+     * @Groups("put")
      */
-    private $publishedAt;
+    private $modifiedAt = null;
 
     /**
-     * @ORM\Column(type="datetime")
+     * @var ?\DateTimeInterface
+     *
+     * @ORM\Column(type="datetime", nullable=true)
+     * @Groups({"get_content", "put"})
      */
-    private $depublishedAt;
+    private $publishedAt = null;
 
     /**
-     * @var Field[]|ArrayCollection
+     * @var ?\DateTimeInterface
+     *
+     * @ORM\Column(type="datetime", nullable=true)
+     * @Groups("put")
+     */
+    private $depublishedAt = null;
+
+    /**
+     * @var Collection|Field[]
+     *
+     * @Groups({"put"})
+     * @MaxDepth(1)
      * @ORM\OneToMany(
      *     targetEntity="Bolt\Entity\Field",
      *     mappedBy="content",
@@ -90,28 +121,15 @@ class Content
      */
     private $fields;
 
-    /**
-     * @var ContentType
-     */
+    /** @var ?ContentType */
     private $contentTypeDefinition;
 
-    /** @var UrlGeneratorInterface */
-    private $urlGenerator;
-
-    /** @var Config */
-    private $config;
-
     /**
-     * Set the "Magic properties for automagic population in the API.
-     */
-    public $magictitle;
-    public $magicexcerpt;
-    public $magicimage;
-    public $magiclink;
-    public $magiceditlink;
-
-    /**
-     * @ORM\ManyToMany(targetEntity="Bolt\Entity\Taxonomy", mappedBy="content")
+     * @var Collection|Taxonomy[]
+     * @Groups({"get_content", "put"})
+     * @MaxDepth(1)
+     *
+     * @ORM\ManyToMany(targetEntity="Bolt\Entity\Taxonomy", mappedBy="content", cascade={"persist"})
      * @ORM\JoinTable(name="bolt_taxonomy_content")
      */
     private $taxonomies;
@@ -119,9 +137,6 @@ class Content
     public function __construct()
     {
         $this->createdAt = new \DateTime();
-        $this->modifiedAt = new \DateTime();
-        $this->publishedAt = new \DateTime();
-        $this->depublishedAt = new \DateTime();
         $this->fields = new ArrayCollection();
         $this->taxonomies = new ArrayCollection();
     }
@@ -132,54 +147,75 @@ class Content
     }
 
     /**
-     * @param Config $config
+     * @see: Bolt\EventListener\ContentListener
      */
-    public function setConfig(Config $config)
+    public function setDefinitionFromContentTypesConfig(LaravelCollection $contentTypesConfig): void
     {
-        $this->config = $config;
-
-        $this->contentTypeDefinition = ContentTypeFactory::get($this->contentType, $config->get('contenttypes'));
+        $this->contentTypeDefinition = ContentType::factory($this->contentType, $contentTypesConfig);
     }
 
-    public function getConfig(): Config
-    {
-        return $this->config;
-    }
-
-    /**
-     * @param UrlGeneratorInterface $urlGenerator
-     */
-    public function setUrlGenerator(UrlGeneratorInterface $urlGenerator)
-    {
-        $this->urlGenerator = $urlGenerator;
-    }
-
-    public function getDefinition()
+    public function getDefinition(): ?ContentType
     {
         return $this->contentTypeDefinition;
     }
 
-    /**
-     * @return string
-     */
-    public function getSlug(): string
+    public function getSummary(): array
     {
-        return  (string) $this->get('slug');
+        if ($this->getDefinition() === null) {
+            return [];
+        }
+
+        return [
+            'id' => $this->getid(),
+            'contentType' => $this->getDefinition()->get('slug'),
+            'slug' => $this->getSlug(),
+            'title' => $this->magicTitle(),
+            'excerpt' => $this->magicExcerpt(200, false),
+            'image' => $this->magicImage(),
+            'link' => $this->magicLink(),
+            'editLink' => $this->magicEditLink(),
+            'author' => [
+                'id' => $this->getAuthor()->getid(),
+                'displayName' => $this->getAuthor()->getDisplayName(),
+                'username' => $this->getAuthor()->getusername(),
+                'email' => $this->getAuthor()->getemail(),
+                'roles' => $this->getAuthor()->getroles(),
+            ],
+            'status' => $this->getStatus(),
+            'icon' => $this->getDefinition()->get('icon_one'),
+            'createdAt' => $this->getCreatedAt(),
+            'modifiedAt' => $this->getModifiedAt(),
+            'publishedAt' => $this->getPublishedAt(),
+            'depublishedAt' => $this->getDepublishedAt(),
+        ];
     }
 
-    public function getContenttype(): ?string
+    public function getSlug(): string
+    {
+        return $this->get('slug')->__toString();
+    }
+
+    public function getContentType(): ?string
     {
         return $this->contentType;
     }
 
-    public function setContenttype(string $contenttype): self
+    public function setContentType(string $contentType): self
     {
-        $this->contentType = $contenttype;
+        $this->contentType = $contentType;
 
         return $this;
     }
 
     public function getAuthor(): User
+    {
+        return $this->author;
+    }
+
+    /**
+     * @deprecated Backward-compatible alias for `getAuthor`
+     */
+    public function geUser(): User
     {
         return $this->author;
     }
@@ -191,17 +227,23 @@ class Content
 
     public function getStatus(): ?string
     {
+        if (Statuses::isValid($this->status) === false) {
+            $this->status = $this->getDefinition()->get('default_status');
+        }
+
         return $this->status;
     }
 
     public function setStatus(string $status): self
     {
-        $this->status = $status;
+        if (Statuses::isValid($status)) {
+            $this->status = $status;
+        }
 
         return $this;
     }
 
-    public function getCreatedAt(): ?\DateTimeInterface
+    public function getCreatedAt(): \DateTimeInterface
     {
         return $this->createdAt;
     }
@@ -218,7 +260,7 @@ class Content
         return $this->modifiedAt;
     }
 
-    public function setModifiedAt(\DateTimeInterface $modifiedAt): self
+    public function setModifiedAt(?\DateTimeInterface $modifiedAt): self
     {
         $this->modifiedAt = $modifiedAt;
 
@@ -242,7 +284,7 @@ class Content
         return $this->depublishedAt;
     }
 
-    public function setDepublishedAt(\DateTimeInterface $depublishedAt): self
+    public function setDepublishedAt(?\DateTimeInterface $depublishedAt): self
     {
         $this->depublishedAt = $depublishedAt;
 
@@ -250,11 +292,43 @@ class Content
     }
 
     /**
-     * @return Collection|Field[]
+     * @return Field[]|Collection
      */
     public function getFields(): Collection
     {
         return $this->fields;
+    }
+
+    /**
+     * @Groups("get_content")
+     */
+    public function getFieldValues(): array
+    {
+        $fieldValues = [];
+        foreach ($this->getFields() as $field) {
+            $fieldValues[$field->getName()] = $field->getFieldValue();
+        }
+
+        return $fieldValues;
+    }
+
+    public function getValue(string $fieldName): ?array
+    {
+        $field = $this->getField($fieldName);
+        return $field ? $field->getValue() : null;
+    }
+
+    /**
+     * @Groups("get_content")
+     */
+    public function getAuthorName(): string
+    {
+        return $this->getAuthor()->getDisplayName();
+    }
+
+    public function hasField(string $name): bool
+    {
+        return collect($this->fields)->contains('name', $name);
     }
 
     public function getField(string $name): ?Field
@@ -264,7 +338,7 @@ class Content
 
     public function addField(Field $field): self
     {
-        if (!$this->fields->contains($field)) {
+        if ($this->fields->contains($field) === false) {
             $this->fields[] = $field;
             $field->setContent($this);
         }
@@ -285,19 +359,16 @@ class Content
         return $this;
     }
 
-    /**
-     * @return array
-     */
     public function getStatuses(): array
     {
-        return self::STATUSES;
+        return Statuses::all();
     }
 
-    public function getStatusOptions()
+    public function getStatusOptions(): array
     {
         $options = [];
 
-        foreach (self::STATUSES as $option) {
+        foreach (Statuses::all() as $option) {
             $options[] = [
                 'key' => $option,
                 'value' => ucwords($option),
@@ -311,14 +382,22 @@ class Content
     /**
      * @return Collection|Taxonomy[]
      */
-    public function getTaxonomies(): Collection
+    public function getTaxonomies(?string $type = null): Collection
     {
+        if ($type) {
+            return $this->taxonomies->filter(
+                function (Taxonomy $taxonomy) use ($type) {
+                    return $taxonomy->getType() === $type;
+                }
+            );
+        }
+
         return $this->taxonomies;
     }
 
     public function addTaxonomy(Taxonomy $taxonomy): self
     {
-        if (!$this->taxonomies->contains($taxonomy)) {
+        if ($this->taxonomies->contains($taxonomy) === false) {
             $this->taxonomies[] = $taxonomy;
             $taxonomy->addContent($this);
         }
@@ -334,5 +413,11 @@ class Content
         }
 
         return $this;
+    }
+
+    public function related(): array
+    {
+        // @todo See Github issue https://github.com/bolt/four/issues/163
+        return [];
     }
 }
