@@ -6,6 +6,7 @@ namespace Bolt\Entity\Field;
 
 use Bolt\Entity\Field;
 use Doctrine\ORM\Mapping as ORM;
+use League\Glide\Urls\UrlBuilderFactory;
 
 /**
  * @ORM\Entity
@@ -14,10 +15,28 @@ class ImageField extends Field
 {
     public function __toString(): string
     {
-        $config = $this->getContent()->getConfig();
+        return $this->getPath();
+    }
 
-        $path = $config->getPath('files', false, $this->get('filename'));
+    public function getValue(): ?array
+    {
+        $value = parent::getValue();
 
-        return $path;
+        // Generate a URL
+        $value['path'] = $this->getPath();
+
+        return $value;
+    }
+
+    private function getPath(): string
+    {
+        $urlBuilder = UrlBuilderFactory::create('/thumbs/');
+
+        // @todo those dimensions shouldn't be hardcoded here
+        return $urlBuilder->getUrl($this->get('filename'), [
+            'w' => 240,
+            'h' => 160,
+            'area' => 'files',
+        ]);
     }
 }
