@@ -5,12 +5,18 @@ declare(strict_types=1);
 namespace Bolt\Entity;
 
 use ApiPlatform\Core\Annotation\ApiResource;
+use Bolt\Utils\Str;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
+use Symfony\Component\Serializer\Annotation\Groups;
 
 /**
- * @ApiResource()
+ * @ApiResource(
+ *     normalizationContext={"groups"={"get_content", "public"}},
+ *     collectionOperations={"get"},
+ *     itemOperations={"get"}
+ * )
  * @ORM\Entity(repositoryClass="Bolt\Repository\TaxonomyRepository")
  * @ORM\Table(name="bolt_taxonomy")
  */
@@ -20,6 +26,7 @@ class Taxonomy
      * @ORM\Id()
      * @ORM\GeneratedValue()
      * @ORM\Column(type="integer")
+     * @Groups("public")
      */
     private $id;
 
@@ -31,27 +38,46 @@ class Taxonomy
 
     /**
      * @ORM\Column(type="string", length=191)
+     * @Groups({"get_content", "public"})
      */
     private $type;
 
     /**
      * @ORM\Column(type="string", length=191)
+     * @Groups({"get_content", "public"})
      */
     private $slug;
 
     /**
      * @ORM\Column(type="string", length=191)
+     * @Groups({"get_content", "public"})
      */
     private $name;
 
     /**
      * @ORM\Column(type="integer")
+     * @Groups("public")
      */
-    private $sortorder;
+    private $sortorder = 0;
 
     public function __construct()
     {
         $this->content = new ArrayCollection();
+    }
+
+    /**
+     * @return Taxonomy
+     */
+    public static function factory(string $type, string $slug, ?string $name = null, int $sortorder = 0): self
+    {
+        $taxonomy = new self();
+
+        $taxonomy->setType($type);
+        $taxonomy->setSlug($slug);
+        $taxonomy->setName($name ?: $slug);
+        $taxonomy->setSortorder($sortorder);
+
+        return $taxonomy;
     }
 
     public function getId(): ?int
@@ -69,7 +95,7 @@ class Taxonomy
 
     public function addContent(Content $content): self
     {
-        if (!$this->content->contains($content)) {
+        if (! $this->content->contains($content)) {
             $this->content[] = $content;
         }
 
@@ -104,7 +130,7 @@ class Taxonomy
 
     public function setSlug(string $slug): self
     {
-        $this->slug = $slug;
+        $this->slug = Str::slug($slug);
 
         return $this;
     }
