@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Bolt\EventListener;
 
 use Bolt\Configuration\Config;
+use Bolt\Entity\Content;
 use Doctrine\ORM\Event\LifecycleEventArgs;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 
@@ -22,12 +23,13 @@ class ContentListener
         $this->urlGenerator = $urlGenerator;
     }
 
-    public function postLoad(LifecycleEventArgs $args)
+    public function postLoad(LifecycleEventArgs $args): void
     {
+        /** @var Content $entity */
         $entity = $args->getEntity();
 
-        if (method_exists($entity, 'setConfig')) {
-            $entity->setConfig($this->config);
+        if (method_exists($entity, 'setDefinitionFromContentTypesConfig')) {
+            $entity->setDefinitionFromContentTypesConfig($this->config->get('contenttypes'));
         }
         if (method_exists($entity, 'setUrlGenerator')) {
             $entity->setUrlGenerator($this->urlGenerator);
