@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Bolt\EventSubscriber;
 
-use Bolt\Twig\Extension\SourceCodeExtension;
+use Bolt\Twig\SourceCodeExtension;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpKernel\Event\FilterControllerEvent;
 use Symfony\Component\HttpKernel\KernelEvents;
