@@ -6,11 +6,12 @@ namespace Bolt\Repository;
 
 use Bolt\Entity\Field;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
+use Doctrine\ORM\QueryBuilder;
 use Symfony\Bridge\Doctrine\RegistryInterface;
 
 /**
- * @method Field|null find($id, $lockMode = null, $lockVersion = null)
- * @method Field|null findOneBy(array $criteria, array $orderBy = null)
+ * @method (Field | null) find($id, $lockMode = null, $lockVersion=null)
+ * @method (Field | null) findOneBy(array $criteria, array $orderBy=null)
  * @method Field[]    findAll()
  * @method Field[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
  */
@@ -21,7 +22,7 @@ class FieldRepository extends ServiceEntityRepository
         parent::__construct($registry, Field::class);
     }
 
-    private function getQueryBuilder(QueryBuilder $qb = null)
+    private function getQueryBuilder(?QueryBuilder $qb = null)
     {
         return $qb ?: $this->createQueryBuilder('field');
     }
@@ -32,8 +33,7 @@ class FieldRepository extends ServiceEntityRepository
             ->andWhere('field.value = :slug')
             ->setParameter('slug', \GuzzleHttp\json_encode([$slug]))
             ->getQuery()
-            ->getOneOrNullResult()
-            ;
+            ->getOneOrNullResult();
     }
 
 //    /**
@@ -51,15 +51,13 @@ class FieldRepository extends ServiceEntityRepository
             ->getResult()
         ;
     }
-    */
-
+     */
     public function findOneBySomeField($value): ?Field
     {
         return $this->createQueryBuilder('field')
             ->andWhere('field.exampleField = :val')
             ->setParameter('val', $value)
             ->getQuery()
-            ->getOneOrNullResult()
-        ;
+            ->getOneOrNullResult();
     }
 }
